src/Tests/TestsChargeOver.Wrapper/Services/TermsServiceTests.cs
using System;
using ChargeOver.Wrapper.Models;
using ChargeOver.Wrapper.Services;
using NUnit.Framework;

namespace TestsChargeOver.Wrapper.Services
{
	[TestFixture]
	public sealed class TermsServiceTests
	{
		private TermsService Sut{get;set;}

		[SetUp]
		public void SetUp()
		{
			Sut = new TermsService(new ChargeOverApiProvider(ChargeOverAPIConfiguration.Config));
		}

		[Test]
		public void should_call_ListTerms()
		{
			//arrange
			//act
			var actual = Sut.ListTerms();
			//assert
			Assert.AreEqual(200, actual.Code);
			Assert.IsEmpty(actual.Message);
			Assert.AreEqual("OK", actual.Status);
		}
	}
}
